Rename highlight-next/src/util/withHighligh.ts
