+++ sdk/highlight-next/src/server.ts
@@ -1,3 +1,7 @@
 export { registerHighlight } from './util/registerHighlight'
-export { H, Highlight, HighlightEnv } from './util/withHighlight'
+export {
+	H,
+	Highlight,
+	HighlightEnv,
+} from './util/with-highlight-nodejs-page-router'
 export { withHighlightConfig } from './util/withHighlightConfig'
