+++ sdk/highlight-next/src/util/withHighlight.test.ts
@@ -1,5 +1,5 @@
 import { NextApiHandler } from 'next'
-import { Highlight } from './withHighlight.js'
+import { Highlight } from './with-highlight-nodejs-page-router.js'
 
 import { createNextApiHandler } from '@trpc/server/adapters/next'
 import * as trpc from '@trpc/server'
